# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para la gestión de ordenes y clientes.",
        "title": "ITLAFINAL API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/customers": {
            "get": {
                "description": "Recupera la lista de todos los clientes registrados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Obtener todos los clientes",
                "responses": {
                    "200": {
                        "description": "Lista de clientes obtenida con éxito"
                    }
                }
            },
            "post": {
                "description": "Registra un nuevo cliente en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Crear un cliente",
                "parameters": [
                    {
                        "description": "Customer data",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCustomerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cliente creado con éxito"
                    }
                }
            }
        },
        "/customers/{id}": {
            "get": {
                "description": "Recupera los detalles de un cliente específico utilizando su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Obtener un cliente por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente obtenido con éxito"
                    }
                }
            }
        },
        "/orders": {
            "get": {
                "description": "Recupera la lista de todas las órdenes registradas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Obtener todas las órdenes",
                "responses": {
                    "200": {
                        "description": "Lista de órdenes obtenida con éxito"
                    }
                }
            },
            "post": {
                "description": "Registra una nueva orden en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Crear una orden",
                "parameters": [
                    {
                        "description": "Order data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Orden creada con éxito"
                    }
                }
            }
        },
        "/orders/{id}": {
            "put": {
                "description": "Actualiza el estado de una orden existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Actualizar el estado de una orden",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la orden",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevo estado de la orden",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado actualizado con éxito"
                    }
                }
            },
            "delete": {
                "description": "Elimina una orden existente de la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Eliminar una orden",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la orden",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orden eliminada con éxito"
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateCustomerRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "phone"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "dto.CreateOrderRequest": {
            "type": "object",
            "required": [
                "customer_id",
                "pieces_count",
                "service_type"
            ],
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "pieces_count": {
                    "type": "integer",
                    "minimum": 1
                },
                "service_type": {
                    "type": "string"
                }
            }
        }
    }
}